requirements.txt
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
inflection==0.5.1
msgpack==1.0.7
ndg-httpsclient==0.5.1
packaging==23.2
pusher==3.3.2
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
whitenoise==6.6.0
